+++ app/src/main/java/cityfreqs/com/pilfershushjammer/ui/InspectorFragment.java
@@ -69,7 +69,6 @@ public class InspectorFragment extends Fragment {
         return view;
     }
 
-
     /**********************************************************************************************/
 
     private void populateElements() {